Create a second JS bundle file
