+++ app/views/layouts/application.html.erb
@@ -17,7 +17,8 @@
     <meta property="og:site_name" content="<%= meta_title %>" />
 
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
-    <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
+    <%= javascript_pack_tag 'bundle1', 'data-turbolinks-track': 'reload' %>
+    <%= javascript_pack_tag 'bundle2', 'data-turbolinks-track': 'reload' %>
   </head>
 
   <body>
